double dialog fixed

+++ src/main/java/cz/muni/fi/pv168/project/business/service/crud/RecipeCrudService.java
@@ -6,7 +6,9 @@ import cz.muni.fi.pv168.project.business.model.Recipe;
 import cz.muni.fi.pv168.project.business.repository.Repository;
 import cz.muni.fi.pv168.project.business.service.validation.RecipeValidator;
 import cz.muni.fi.pv168.project.business.service.validation.ValidationResult;
+import cz.muni.fi.pv168.project.ui.action.TabbedPanelContext;
 
+import javax.swing.*;
 import java.util.List;
 
 public class RecipeCrudService implements CrudService<Recipe> {
@@ -50,7 +52,7 @@ public class RecipeCrudService implements CrudService<Recipe> {
 //            Creating added ingredient via their service
             newEntity.getAddedIngredients().forEach(addedIngredientCrudService::create);
 //            adding recipe to usedIn list in category, later should decline deleting this category
-            newEntity.getCategory().addRecipe(newEntity);
+            //newEntity.getCategory().addRecipe(newEntity);
         }
 
         return validationResult;
@@ -68,6 +70,14 @@ public class RecipeCrudService implements CrudService<Recipe> {
 
     @Override
     public ValidationResult deleteByGuid(String guid, boolean userAgreed) {
+        int confirm = JOptionPane.showOptionDialog(TabbedPanelContext.getActiveTable(),
+                "Confirm",
+                "Delete confirmation",
+                JOptionPane.YES_NO_OPTION,JOptionPane.PLAIN_MESSAGE,
+                null,null,null);
+        if ( confirm != JOptionPane.OK_OPTION) {
+            return ValidationResult.failed("denied");
+        }
         recipeRepository.deleteByGuid(guid);
         return ValidationResult.success();
     }
